+++ sdk/ai/azure-ai-generative/azure/ai/generative/index/_tasks/generate_qa.py
@@ -72,7 +72,11 @@ def get_model_config(llm_config: Dict[str, str], openai_api_type: str, openai_ap
 
     # Only add base and version if using AOAI
     if model_config["api_type"] == "azure":
-        openai.api_base = model_config["api_base"]
+        # openai.api_base is replaced by openai.base_url in openai 1.x
+        if hasattr(openai, "api_base"):
+            openai.api_base = model_config["api_base"]
+        else:
+            openai.base_url = model_config["api_base"]
         openai.api_version = model_config["api_version"]
     return model_config