+++ sdk/appconfiguration/azure-appconfiguration-provider/tests/testcase.py
@@ -22,6 +22,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
         secret_resolver=None,
         key_vault_options=None,
         on_refresh_success=None,
+        feature_flag_enabled=False,
     ):
         cred = self.get_credential(AzureAppConfigurationClient)
         client = AzureAppConfigurationClient(appconfiguration_endpoint_string, cred)
@@ -39,6 +40,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
                 user_agent="SDK/Integration",
                 keyvault_credential=keyvault_cred,
                 on_refresh_success=on_refresh_success,
+                feature_flag_enabled=feature_flag_enabled,
             )
         if key_vault_options:
             if not key_vault_options.secret_resolver:
@@ -53,6 +55,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
                 user_agent="SDK/Integration",
                 key_vault_options=key_vault_options,
                 on_refresh_success=on_refresh_success,
+                feature_flag_enabled=feature_flag_enabled,
             )
         return load(
             credential=cred,
@@ -64,6 +67,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
             user_agent="SDK/Integration",
             secret_resolver=secret_resolver,
             on_refresh_success=on_refresh_success,
+            feature_flag_enabled=feature_flag_enabled,
         )
 
     def create_client(
@@ -77,6 +81,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
         secret_resolver=None,
         key_vault_options=None,
         on_refresh_success=None,
+        feature_flag_enabled=False,
     ):
         client = AzureAppConfigurationClient.from_connection_string(appconfiguration_connection_string)
         setup_configs(client, keyvault_secret_url)
@@ -91,6 +96,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
                 user_agent="SDK/Integration",
                 keyvault_credential=self.get_credential(AzureAppConfigurationClient),
                 on_refresh_success=on_refresh_success,
+                feature_flag_enabled=feature_flag_enabled,
             )
         if key_vault_options:
             if not key_vault_options.secret_resolver:
@@ -106,6 +112,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
                 user_agent="SDK/Integration",
                 key_vault_options=key_vault_options,
                 on_refresh_success=on_refresh_success,
+                feature_flag_enabled=feature_flag_enabled,
             )
         return load(
             connection_string=appconfiguration_connection_string,
@@ -116,6 +123,7 @@ class AppConfigTestCase(AzureRecordedTestCase):
             user_agent="SDK/Integration",
             secret_resolver=secret_resolver,
             on_refresh_success=on_refresh_success,
+            feature_flag_enabled=feature_flag_enabled,
         )