+++ sdk/documentintelligence/azure-ai-documentintelligence/samples/sample_analyze_documents_output_in_markdown.py
@@ -41,7 +41,7 @@ def analyze_documents_output_in_markdown():
     )
     result = poller.result()
 
-    print(f"Here's the full content in format{result.content_format}:\n")
+    print(f"Here's the full content in format {result.content_format}:\n")
     print(result.content)
     # [END analyze_documents_output_in_markdown]