+++ sdk/documentintelligence/azure-ai-documentintelligence/tests/test_dac_analyze_layout_async.py
@@ -7,6 +7,7 @@
 import pytest
 import functools
 from devtools_testutils.aio import recorded_by_proxy_async
+from azure.core.credentials import AzureKeyCredential
 from azure.ai.documentintelligence.aio import DocumentIntelligenceClient
 from azure.ai.documentintelligence.models import DocumentAnalysisFeature, AnalyzeDocumentRequest
 from asynctestcase import AsyncDocumentIntelligenceTest
@@ -144,3 +145,31 @@ class TestDACAnalyzeLayoutAsync(AsyncDocumentIntelligenceTest):
         assert layout.pages[0].barcodes[0].kind == "Code39"
         assert layout.pages[0].barcodes[0].polygon
         assert layout.pages[0].barcodes[0].confidence > 0.8
+    
+    @skip_flaky_test
+    @DocumentIntelligencePreparer()
+    @recorded_by_proxy_async
+    async def test_polling_interval(self, documentintelligence_endpoint, documentintelligence_api_key, **kwargs):
+        client = DocumentIntelligenceClient(
+            documentintelligence_endpoint, AzureKeyCredential(documentintelligence_api_key)
+        )
+        assert client._config.polling_interval ==  5
+        
+        client = DocumentIntelligenceClient(
+            documentintelligence_endpoint, AzureKeyCredential(documentintelligence_api_key), polling_interval=7
+        )
+        assert client._config.polling_interval ==  7
+        async with client:
+            poller = await client.begin_analyze_document(
+                "prebuilt-receipt",
+                AnalyzeDocumentRequest(url_source=self.receipt_url_jpg),
+                polling_interval=6
+            )
+            await poller.wait()
+            assert poller._polling_method._timeout ==  6
+            poller2 = await client.begin_analyze_document(
+                "prebuilt-receipt",
+                AnalyzeDocumentRequest(url_source=self.receipt_url_jpg),
+            )
+            await poller2.wait()
+            assert poller2._polling_method._timeout ==  7  # goes back to client default