+++ sdk/ml/azure-ai-ml/tests/workspace/unittests/test_workspace_connection_entity.py
@@ -8,6 +8,7 @@ from azure.ai.ml._restclient.v2023_06_01_preview.models import ConnectionAuthTyp
 from azure.ai.ml._utils.utils import camel_to_snake
 from azure.ai.ml.entities import WorkspaceConnection, AzureOpenAIWorkspaceConnection
 from azure.ai.ml.entities._credentials import PatTokenConfiguration
+from azure.ai.ml.constants._common import WorkspaceConnectionTypes
 
 
 @pytest.mark.unittest
@@ -171,6 +172,15 @@ class TestWorkspaceConnectionEntity:
         assert ws_connection.kind == "some_kind"
         assert ws_connection.api_version == "dummy"
 
+        ws_connection = load_workspace_connection(source="./tests/test_configs/workspace_connection/custom.yaml")
+
+        assert ws_connection.type == camel_to_snake(WorkspaceConnectionTypes.CUSTOM)
+        assert ws_connection.credentials.type == camel_to_snake(ConnectionAuthType.API_KEY)
+        assert ws_connection.credentials.key == "4444"
+        assert ws_connection.name == "test_ws_conn_custom_keys"
+        assert ws_connection.target == "my_endpoint"
+        assert ws_connection.tags["one"] == "two"
+
     def test_ws_conn_rest_conversion(self):
         ws_connection = load_workspace_connection(source="./tests/test_configs/workspace_connection/open_ai.yaml")
         rest_conn = ws_connection._to_rest_object()
@@ -182,3 +192,4 @@ class TestWorkspaceConnectionEntity:
         assert ws_connection.tags["hello"] == new_ws_conn.tags["hello"]
         assert ws_connection.api_version == new_ws_conn.api_version
         assert ws_connection.api_type == new_ws_conn.api_type
+        assert ws_connection.is_shared