+++ sdk/documentintelligence/azure-ai-documentintelligence/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Changed the default polling interval from 30s to 5s.
+
 ## 1.0.0b1 (2023-11-17)
 
 This is the first preview of the `azure-ai-documentintelligence` package, targeting API version `2023-10-31-preview` of the Document Intelligence service(formerly known as Form Recognizer).